+++ cli/docs/changelog.md
@@ -1,5 +1,14 @@
 # TBD
 
+### Breaking Changes
+* Removed magical constant `IPADDR` that would get replaced with the IP address.
+  * Added a flag for the user to supply a placeholder using `ip-address-placeholder` in `kurtosis enclave add`. Occurrences of this flag in entrypoint args, env vars & cmd args will get replaced with the private IP address inside the container.
+  * Users should use this new flag to do any IP address replacements.
+
+### Changes
+* Removed magical constant `IPADDR` that would get replaced with the IP address.
+* Added a flag for the user to supply a placeholder using `ip-address-placeholder` in `kurtosis enclave add`. Occurrences of this flag in entrypoint args, env vars & cmd args will get replaced with the private IP address inside the container.
+
 ### Features
 * Added `rendertemplate` to `kurtosis files` which allows you to upload a Golang text/template with data provided as a JSON file.